Move some code into a separate file and add docs
